requirements.txt
tweepy==4.10.0
transformers==4.21.1
pandas==1.4.3
numpy==1.23.1
matplotlib==3.5.2
plotly==5.10.0
python-dotenv==0.20.0
